LPS-54053 OSGi modules can't share the same packages
