+++ modules/apps/wiki/wiki-test/test/integration/com/liferay/wiki/service/test/WikiPageLocalServiceTest.java
@@ -49,6 +49,7 @@ import com.liferay.wiki.exception.NoSuchPageResourceException;
 import com.liferay.wiki.exception.NodeChangeException;
 import com.liferay.wiki.model.WikiNode;
 import com.liferay.wiki.model.WikiPage;
+import com.liferay.wiki.service.WikiPageLocalServiceUtil;
 import com.liferay.wiki.util.test.WikiTestUtil;
 
 import java.util.ArrayList;
