+++ modules/apps/wiki/wiki-test/test/integration/com/liferay/wiki/social/test/WikiActivityInterpreterTest.java
@@ -32,6 +32,8 @@ import com.liferay.portlet.trash.util.TrashUtil;
 import com.liferay.wiki.model.WikiNode;
 import com.liferay.wiki.model.WikiPage;
 import com.liferay.wiki.service.WikiPageLocalServiceUtil;
+import com.liferay.wiki.social.WikiActivityInterpreter;
+import com.liferay.wiki.social.WikiActivityKeys;
 import com.liferay.wiki.util.test.WikiTestUtil;
 
 import org.junit.ClassRule;
