+++ modules/apps/wiki/wiki-test/test/integration/com/liferay/wiki/subscription/test/WikiSubscriptionContainerModelTest.java
@@ -12,20 +12,17 @@
  * details.
  */
 
-package com.liferay.wiki.subscription;
+package com.liferay.wiki.subscription.test;
 
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.kernel.test.rule.Sync;
 import com.liferay.portal.kernel.test.rule.SynchronousMailTestRule;
-import com.liferay.portal.kernel.test.util.RoleTestUtil;
-import com.liferay.portal.model.ResourceConstants;
-import com.liferay.portal.model.RoleConstants;
-import com.liferay.portal.security.permission.ActionKeys;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 import com.liferay.portal.test.rule.MainServletTestRule;
-import com.liferay.portlet.subscriptions.test.BaseSubscriptionBaseModelTestCase;
+import com.liferay.portlet.subscriptions.test.BaseSubscriptionContainerModelTestCase;
 import com.liferay.wiki.model.WikiNode;
 import com.liferay.wiki.model.WikiPage;
+import com.liferay.wiki.service.WikiNodeLocalServiceUtil;
 import com.liferay.wiki.service.WikiPageLocalServiceUtil;
 import com.liferay.wiki.util.test.WikiTestUtil;
 
@@ -39,8 +36,8 @@ import org.junit.Test;
  * @author Roberto Díaz
  */
 @Sync
-public class WikiSubscriptionBaseModelTest
-	extends BaseSubscriptionBaseModelTestCase {
+public class WikiSubscriptionContainerModelTest
+	extends BaseSubscriptionContainerModelTestCase {
 
 	@ClassRule
 	@Rule
@@ -52,7 +49,25 @@ public class WikiSubscriptionBaseModelTest
 	@Ignore
 	@Override
 	@Test
-	public void testSubscriptionBaseModelWhenInRootContainerModel() {
+	public void testSubscriptionContainerModelWhenAddingBaseModelInRootContainerModel() {
+	}
+
+	@Ignore
+	@Override
+	@Test
+	public void testSubscriptionContainerModelWhenAddingBaseModelInSubcontainerModel() {
+	}
+
+	@Ignore
+	@Override
+	@Test
+	public void testSubscriptionContainerModelWhenUpdatingBaseModelInRootContainerModel() {
+	}
+
+	@Ignore
+	@Override
+	@Test
+	public void testSubscriptionContainerModelWhenUpdatingBaseModelInSubcontainerModel() {
 	}
 
 	@Override
@@ -65,32 +80,17 @@ public class WikiSubscriptionBaseModelTest
 
 	@Override
 	protected long addContainerModel(long containerModelId) throws Exception {
-		_node = WikiTestUtil.addNode(group.getGroupId());
+		WikiNode node = WikiTestUtil.addNode(group.getGroupId());
 
-		return _node.getNodeId();
+		return node.getNodeId();
 	}
 
 	@Override
-	protected void addSubscriptionBaseModel(long baseModelId) throws Exception {
-		WikiPage page = WikiPageLocalServiceUtil.getPage(baseModelId);
-
-		WikiPageLocalServiceUtil.subscribePage(
-			user.getUserId(), page.getNodeId(), page.getTitle());
-	}
-
-	@Override
-	protected void removeContainerModelResourceViewPermission()
+	protected void addSubscriptionContainerModel(long containerModelId)
 		throws Exception {
 
-		RoleTestUtil.removeResourcePermission(
-			RoleConstants.GUEST, WikiNode.class.getName(),
-			ResourceConstants.SCOPE_INDIVIDUAL,
-			String.valueOf(_node.getNodeId()), ActionKeys.VIEW);
-
-		RoleTestUtil.removeResourcePermission(
-			RoleConstants.SITE_MEMBER, WikiNode.class.getName(),
-			ResourceConstants.SCOPE_INDIVIDUAL,
-			String.valueOf(_node.getNodeId()), ActionKeys.VIEW);
+		WikiNodeLocalServiceUtil.subscribeNode(
+			user.getUserId(), containerModelId);
 	}
 
 	@Override
@@ -99,7 +99,5 @@ public class WikiSubscriptionBaseModelTest
 
 		WikiTestUtil.updatePage(page);
 	}
-
-	private WikiNode _node;
 
 }
