+++ modules/apps/wiki/wiki-test/test/integration/com/liferay/wiki/subscription/test/WikiSubscriptionLocalizedContentTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.wiki.subscription;
+package com.liferay.wiki.subscription.test;
 
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.kernel.test.rule.Sync;
