+++ modules/apps/wiki/wiki-test/test/integration/com/liferay/wiki/translator/test/ClassicToCreoleTranslatorTest.java
@@ -15,6 +15,7 @@
 package com.liferay.wiki.translator.test;
 
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
+import com.liferay.wiki.translator.ClassicToCreoleTranslator;
 
 import org.junit.Assert;
 import org.junit.ClassRule;
