+++ modules/apps/wiki/wiki-test/test/integration/com/liferay/wiki/verify/test/VerifierTest.java
@@ -19,6 +19,7 @@ import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 import com.liferay.portal.test.rule.MainServletTestRule;
 import com.liferay.portal.verify.VerifyProcess;
 import com.liferay.portal.verify.test.BaseVerifyProcessTestCase;
+import com.liferay.wiki.verify.WikiServiceVerifyProcess;
 
 import org.junit.ClassRule;
 import org.junit.Rule;
