+++ source-formatter.properties
@@ -101,8 +101,8 @@
     # 80 column limit.
     #
     line.length.excludes.files=\
-        modules/apps/wiki/wiki-test/test/integration/com/liferay/wiki/subscription/WikiSubscriptionContainerModelTest.java,\
-        modules/apps/wiki/wiki-test/test/integration/com/liferay/wiki/trash/WikiPageDependentsTrashHandlerTest.java,\
+        modules/apps/wiki/wiki-test/test/integration/com/liferay/wiki/subscription/test/WikiSubscriptionContainerModelTest.java,\
+        modules/apps/wiki/wiki-test/test/integration/com/liferay/wiki/trash/test/WikiPageDependentsTrashHandlerTest.java,\
         modules/core/osgi-service-tracker-map/test/integration/com/liferay/osgi/service/tracker/map/test/ObjectServiceTrackerMapTest.java@431,\
         modules/portal/portal-template-freemarker/src/com/liferay/portal/template/freemarker/configuration/FreeMarkerEngineConfiguration.java@52,\
         modules/util/osgi-util/test/integration/com/liferay/osgi/util/test/ReflectionServiceTrackerTest.java@166,\
